+++ pkg/application/interfaces.go
@@ -47,7 +47,7 @@ type applyBlueprintSpecUseCase interface {
 
 type ecosystemHealthUseCase interface {
 	CheckEcosystemHealth(ctx context.Context, ignoreDoguHealth bool, ignoreComponentHealth bool) (ecosystem.HealthResult, error)
-	WaitForHealthyEcosystem(ctx context.Context) (ecosystem.HealthResult, error)
+	WaitForHealthyEcosystem(ctx context.Context, ignoreDoguHealth bool, ignoreComponentHealth bool) (ecosystem.HealthResult, error)
 }
 
 type selfUpgradeUseCase interface {