+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.4.1] - 2025-04-17
+### Changed
+- [#109] Use retry for upfront health checks
+   - blueprints will now not fail immediately if a dogu or component is not yet healthy, no workaround needed anymore
+
 ## [v2.4.0] - 2025-03-31
 ### Added
 - [#107] Add additional print columns and aliases to CRDs